requirements.txt
fastapi
uvicorn
docker
pymongo
python-dotenv
